latest work on getting micros lifecycle events working
