+++ package.json
@@ -90,7 +90,7 @@
     "moo": "^0.5.0",
     "pg": "^8.6.0",
     "primer": "^10.6.0",
-    "probot": "^9.11.3",
+    "probot": "^9.15.1",
     "rate-limit-redis": "^1.7.0",
     "sequelize": "^5.21.7",
     "sequelize-cli": "^6.2.0",
